requirements.txt
asgiref==3.6.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.52.0
django-cors-headers==3.14.0
django-recaptcha==3.0.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
httplib2==0.21.0
idna==3.4
mysqlclient==2.1.1
oauthlib==3.2.2
Pillow==9.4.0
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pymongo==3.12.3
pyotp==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
Sendinblue==2.0.5.1
sib-api-v3-sdk==7.6.0
six==1.16.0
sqlparse==0.2.4
twilio==7.16.5
tzdata==2022.7
urllib3==1.26.15
validate-email==1.3
